gemini_realtime_impl.py
"""
Gemini Realtime LLMの実装
vision-agentsのRealtime抽象クラスを実装
"""

import os
import asyncio
from typing import Optional
from dotenv import load_dotenv

from vision_agents.core.llm.realtime import Realtime
import google.generativeai as genai

load_dotenv()


class GeminiRealtime(Realtime):
    """
    Gemini Realtime LLMの実装
    Google Generative AIを使用
    """
    
    def __init__(self, fps: int = 5):
        super().__init__(fps=fps)
        
        # Gemini APIキーを環境変数から取得
        api_key = os.getenv("GEMINI_API_KEY")
        if not api_key:
            raise ValueError("GEMINI_API_KEY must be set in environment variables")
        
        # Gemini APIを初期化
        genai.configure(api_key=api_key)
        
        # Realtimeモデルを初期化
        # 注意: 実際のGemini Realtime APIの使用方法に合わせて調整が必要
        self.model = None
        self._connected = False
    
    async def connect(self):
        """Realtime APIに接続"""
        try:
            # Gemini Realtime APIに接続
            # 実際の実装では、適切な接続処理が必要
            # ここでは簡易実装
            self._connected = True
            self.emit("connected")
            logger.info("Gemini Realtime APIに接続しました")
        except Exception as e:
            logger.error(f"接続エラー: {e}")
            raise
    
    async def simple_audio_response(self, text: str):
        """音声応答を送信"""
        if not self._connected:
            await self.connect()
        
        try:
            # テキストを音声に変換して送信
            # 実際の実装では、Gemini Realtime APIを使用
            # ここでは簡易実装
            logger.info(f"音声応答を送信: {text}")
            self.emit("audio_response", text)
        except Exception as e:
            logger.error(f"音声応答エラー: {e}")
            raise
    
    async def close(self):
        """接続を閉じる"""
        try:
            self._connected = False
            # リソースをクリーンアップ
            self.emit("closed")
            logger.info("Gemini Realtime APIの接続を閉じました")
        except Exception as e:
            logger.error(f"閉じるエラー: {e}")
    
    async def watch_video_track(self, track):
        """ビデオトラックを監視"""
        # ビデオトラックの監視処理
        # 実際の実装では、フレームを処理してLLMに送信
        logger.info("ビデオトラックを監視中")


# ロガーの設定
import logging
logger = logging.getLogger(__name__)

