stream_edge_implementation.py
"""
StreamEdgeの代替実装
vision-agentsのプラグインが見つからない場合の代替実装
"""

import os
import asyncio
from typing import Optional
from dotenv import load_dotenv

from vision_agents.core.edge.edge_transport import EdgeTransport
from vision_agents.core.edge.types import User, OutputAudioTrack
from getstream.video.client import VideoClient

load_dotenv()


class StreamEdge(EdgeTransport):
    """
    StreamEdgeの実装
    getstreamパッケージを使用してEdgeTransportを実装
    """
    
    def __init__(self, loop: Optional[asyncio.AbstractEventLoop] = None):
        super().__init__(loop=loop)
        
        # Stream APIキーを環境変数から取得
        api_key = os.getenv("STREAM_API_KEY")
        api_secret = os.getenv("STREAM_API_SECRET")
        
        if not api_key or not api_secret:
            raise ValueError("STREAM_API_KEY and STREAM_API_SECRET must be set in environment variables")
        
        # VideoClientを作成
        # 注意: 実際の実装では、適切な初期化が必要
        self.client = VideoClient(
            api_key=api_key,
            base_url="https://video.stream-io-api.com",
            token=api_secret,  # 実際には適切なトークン生成が必要
            timeout=30,
            stream=None,  # 実際の実装では適切なStreamオブジェクトが必要
        )
    
    async def create_user(self, user: User):
        """ユーザーを作成"""
        # 実際の実装では、Stream APIを使用してユーザーを作成
        pass
    
    def create_audio_track(self) -> OutputAudioTrack:
        """オーディオトラックを作成"""
        # 実際の実装では、適切なOutputAudioTrackを作成
        # これは抽象メソッドなので、実装が必要
        raise NotImplementedError("create_audio_track must be implemented")
    
    async def close(self):
        """接続を閉じる"""
        # 実際の実装では、リソースをクリーンアップ
        pass

