requirements.txt
vision-agents>=0.2.0
flask>=2.3.0
python-dotenv>=1.0.0
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0
pillow>=10.0.0
torch>=2.0.0
torchvision>=0.15.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
requests>=2.31.0
websockets>=11.0.0
aiortc>=1.6.0
av>=10.0.0
pydub>=0.25.0